+++ Dominio/Persona.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Persona
     {
-        public int IdPersona { get; set; }
+        public Int64 IdPersona { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public int Dni { get; set; }
